+++ src/pages/lanternDetail/LanternDetail.tsx
@@ -4,7 +4,7 @@ import { CloseButton } from './components/CloseButton/CloseButton';
 import { createMockData } from './constants/mockData';
 import { useCloseGesture } from './hooks/useCloseGesture';
 import * as styles from './LanternDetail.css';
-import { Toast } from '@/components/common/Toast';
+import { Toast } from '@/components/common/Toast/Toast';
 import { LanternData } from '@/components/lantern/constants';
 import { useHandMark } from '@/components/lantern/hooks/useHandMark';
 import { VideoFeed } from '@/components/lantern/VideoFeed';
